feat: Merge om-socket-server stuff. ...
The original plan once upon a time was to open source om-socket-server
and leave this one closed sourced. That's why om-socket-server exists
in the first place. It was the subset of stuff from _this_ package
that we wanted to open source. I can't even remember the reason
now why we wanted to leave this one closed sourced, but anyway we've
decided to open source this package too, so it makes no sense to split
out the functionality. This commit moves the server stuff back into
this package.